Task-04: split search bar into 3 components. Add mobile design for them
