+++ src/components/ProductsCategories/ProductsCategories.module.css
@@ -1,0 +1,36 @@
+
+.categoriesFilter {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.categoryButton {
+  background-color: var(--background-main);
+  color: var(--button-text-secondary);
+  border-radius: 50px;
+  border: 1px solid var(--button-border-third);
+}
+
+.categoryButtonActive {
+  color: var(--text-color-header);
+  border-radius: 50px;
+  background-color: var(--main-red);
+}
+
+@media screen and (min-width: 720px) {
+
+  .categoriesFilter {
+    display: none;
+  }
+}
+
+@media screen and (min-width: 1024px) {
+
+  .categoriesFilter {
+    display: flex;
+    align-items: center;
+    gap: 20px;
+    margin-right: 60px;
+  }
+}
