+++ src/components/SearchBar/SearchBar.module.css
@@ -1,114 +1,59 @@
 .searchBar {
-  display: none;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  align-items: center;
+  margin-bottom: -20px;
+  margin-top: 30px;
+  gap: 20px
 }
 
-.categoriesFilter {
-  display: none;
+.searchBarWrapper {
+  gap: 20px;
+  display: flex;
+  flex-direction: column;
+}
+
+.searchBarRight {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+}
+
+.categoryButton {
+  background-color: var(--background-main);
+  color: var(--button-text-secondary);
+  border-radius: 50px;
+  border: 1px solid var(--button-border-third);
+}
+
+.categoryButtonActive {
+  color: var(--text-color-header);
+  border-radius: 50px;
+  background-color: var(--main-red);
 }
 
 @media screen and (min-width: 720px) {
 
   .searchBar {
     display: flex;
-    margin-top: 50px;
-    margin-left: 70px;
-    margin-right: 70px;
     justify-content: space-between;
-    height: 40px;
-  }
-
-  .searchForm {
-    display: flex;
-  }
-
-  .searchInput {
-    border-radius: 20px;
-    width: 275px;
-    background: var(--background-main);
-    color: var(--text-color);
-    border: 1px solid var(--button-border-third);
-    text-indent: 20px;
-    font-weight: 100;
-  }
-
-  .searchButton {
-    display: flex;
+    width: 100%;
     align-items: center;
-    position: relative;
-    right: 60px;
-    background: none;
+    margin-bottom: -20px;
+    margin-top: 30px;
+  }
+
+  .searchBarWrapper {
+    display: flex;
+    justify-content: space-between;
+    flex-direction: row;
   }
 
   .searchBarRight {
     display: flex;
-    align-items: center;
-  }
-
-  .categoryButton {
-    background-color: var(--background-main);
-    color: var(--button-text-secondary);
-    border-radius: 50px;
-    border: 1px solid var(--button-border-third);
-  }
-
-  .categoryButtonActive {
-    color: var(--text-color-header);
-    border-radius: 50px;
-    background-color: var(--main-red);
-  }
-
-  .sortOptions {
-    display: flex;
-    align-items: center;
-    color: var(--text-color-header);
-    gap: 10px;
-    width: 270px;
-  }
-
-  .sortSpan {
-    color: var(--text-color-header);
-  }
-
-  .dropdownBtn {
-    -webkit-appearance: none;
-    width: 200px;
-    color: var(--button-text-secondary);
-    border: 1px solid var(--button-border-third);
-    border-radius: 50px;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    float: left;
-    z-index: 1;
-  }
-
-  .dropdownContent {
-    position: absolute;
-    width: 200px;
-    right: 70px;
-    top: 130px;
-    padding-top: 40px;
-    background: var(--button-background-third);
-    border-radius: 20px;
-    border: 1px solid var(--button-border-third);
-  }
-
-  .dropdownOption {
-    display: flex;
-    border-radius: 20px;
-    padding: 12px 16px;
-    width: 100%;
-    color: var(--button-text-secondary);
-  }
-
-  .dropdownOptionActive {
-    display: flex;
-    border-radius: 20px;
-    background: var(--main-red);
-    padding: 12px 16px;
-    width: 100%;
-    color: var(--white);
+    flex-direction: row;
   }
 }
