+++ src/constants/products-categories.constant.ts
@@ -1,0 +1,1 @@
+export const categories = ['Electronics', 'Shoes', 'Clothes'] as const;
