+++ src/constants/sort-options.constant.ts
@@ -1,0 +1,1 @@
+export const sortOptions = ['Price (Low-High)', 'Newest', 'Oldest'] as const;
